+++ NEWS.md
@@ -1,3 +1,4 @@
 # chess 1.0.0.9000
 
+* `install_chess()` must use Python 3.8.
 * Added a `NEWS.md` file to track changes to the package.